requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
Pillow>=9.0.0
tqdm>=4.65.0
wandb>=0.15.0